downloadImage function moved to UIImageView+Extensions.
